+++ CHANGELOG.rst
@@ -16,3 +16,11 @@ Changelog
 2.0.1 (2021-12-03)
 ------------------
 * Made the last error accessible
+
+2.0.2 (2021-12-04)
+------------------
+* Added optional argument `differential_input` to have the PID controller
+  regulate a specific difference with respect to the main argument
+  `current_input`. The specific difference that will be regulated is set by
+  `setpoint`.
+  Contributor: https://github.com/antonverburg.